+++ docs/source/changelog.rst
@@ -6,6 +6,17 @@ v2.0.0 (2022-4-22)
 * Added support for attributing ECFP, MACCS fingerprints, rdkit descriptors and plotting them
 * Example notebooks for new method
 
+
+v1.1.0 (2022-5-2)
+-------------------
+* Made it possible to have model function that only takes one argument instead of SMILES and SELFIES
+
+v1.0.2 (2022-5-2)
+-------------------
+* Switched to bulk Tanimoto to improve speed
+* Tightened chemed api limit
+
+
 v1.0.1 (2022-4-4)
 -------------------
 * Added quiet mode